Script to load mongo historian data into hourly, daily, and monthly rolled_up collections. Provide a start and start time in main method. add a temporary index on ts field in data collection to help with performance of this script
